requirements.txt
requests==2.32.3
python-decouple==3.8
python-telegram-bot~=13.0